+++ src/main/java/com/leinardi/pycharm/pylint/plapi/PylintRunner.java
@@ -39,6 +39,7 @@ import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.Moshi;
 import com.squareup.moshi.Types;
 import okio.Okio;
+import org.jdesktop.swingx.util.OS;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.BufferedReader;
@@ -56,15 +57,18 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 
 public class PylintRunner {
     public static final String PYLINT_PACKAGE_NAME = "pylint";
+    private static final String PYLINT_EXECUTABLE_NAME = PYLINT_PACKAGE_NAME + (OS.isWindows() ? ".exe" : "");
     private static final Logger LOG = com.intellij.openapi.diagnostic.Logger.getInstance(PylintRunner.class);
     private static final String ENV_KEY_VIRTUAL_ENV = "VIRTUAL_ENV";
     private static final String ENV_KEY_PATH = "PATH";
     private static final String ENV_KEY_PYTHONHOME = "PYTHONHOME";
+    private static final String WHICH_EXECUTABLE_NAME = OS.isWindows() ? "where" : "which";
 
     private PylintRunner() {
     }
@@ -76,6 +80,7 @@ public class PylintRunner {
         }
         VirtualFile pylintFile = LocalFileSystem.getInstance().findFileByPath(pylintPath);
         if (pylintFile == null || !pylintFile.exists()) {
+            LOG.error("Error while checking Pylint path " + pylintPath + ": null or not exists");
             return false;
         }
         GeneralCommandLine cmd = getPylintCommandLine(project, pylintPath);
@@ -85,8 +90,24 @@ public class PylintRunner {
         try {
             process = cmd.createProcess();
             process.waitFor();
-            return process.exitValue() == 0;
+            String error = new BufferedReader(new InputStreamReader(process.getErrorStream(), UTF_8))
+                    .lines().collect(Collectors.joining("\n"));
+            if (!StringUtil.isEmpty(error)) {
+                LOG.error("Error while checking Pylint path: " + error);
+            }
+            String output = new BufferedReader(new InputStreamReader(process.getInputStream(), UTF_8))
+                    .lines().collect(Collectors.joining("\n"));
+            if (!StringUtil.isEmpty(output)) {
+                LOG.debug("Pylint path check output: " + output);
+            }
+            if (process.exitValue() != 0) {
+                LOG.error("Pylint path check process.exitValue: " + process.exitValue());
+                return false;
+            } else {
+                return true;
+            }
         } catch (ExecutionException | InterruptedException e) {
+            LOG.error("Error while checking Pylint path", e);
             return false;
         }
     }
@@ -111,7 +132,7 @@ public class PylintRunner {
         VirtualFile interpreterFile = getInterpreterFile(project);
         if (isVenv(interpreterFile)) {
             VirtualFile pylintFile = LocalFileSystem.getInstance()
-                    .findFileByPath(interpreterFile.getParent().getPath() + File.separator + PYLINT_PACKAGE_NAME);
+                    .findFileByPath(interpreterFile.getParent().getPath() + File.separator + PYLINT_EXECUTABLE_NAME);
             if (pylintFile != null && pylintFile.exists()) {
                 return pylintFile.getPath();
             }
@@ -172,8 +193,8 @@ public class PylintRunner {
     }
 
     public static String detectSystemPylintPath() {
-        GeneralCommandLine cmd = new GeneralCommandLine("which");
-        cmd.addParameter(PYLINT_PACKAGE_NAME);
+        GeneralCommandLine cmd = new GeneralCommandLine(WHICH_EXECUTABLE_NAME);
+        cmd.addParameter(PYLINT_EXECUTABLE_NAME);
         final Process process;
         try {
             process = cmd.createProcess();
@@ -182,9 +203,16 @@ public class PylintRunner {
                     .lines()
                     .findFirst();
             process.waitFor();
+            String error = new BufferedReader(new InputStreamReader(process.getErrorStream(), UTF_8))
+                    .lines().collect(Collectors.joining("\n"));
+            if (!StringUtil.isEmpty(error)) {
+                LOG.error("Error while detecting Pylint path: " + error);
+            }
             if (process.exitValue() != 0 || !path.isPresent()) {
+                LOG.error("Pylint path detect process.exitValue: " + process.exitValue());
                 return "";
             }
+            LOG.info("Detected Pylint path: " + path.get());
             return path.get();
         } catch (ExecutionException | InterruptedException e) {
             return "";